Data.Label.Point

Description

The Point data type which generalizes the different lenses and forms the basis for vertical composition using the Applicative type class.

The point data type that generalizes lens.

data Point cat g i f o

Abstract Point datatype. The getter and modifier operations work in some category. The type of the value pointed to might change, thereby changing the type of the outer structure.

Constructors

Point (cat f o) (cat (cat o i, f) g)

get :: Point cat g i f o -> cat f o

Get the getter category from a Point.

modify :: Point cat g i f o -> cat (cat o i, f) g

Get the modifier category from a Point.

set :: Arrow arr => Point arr g i f o -> arr (i, f) g

Get the setter category from a Point.

identity :: ArrowApply arr => Point arr f f o o

Identity Point. Cannot change the type.

compose :: ArrowApply cat => Point cat t i b o -> Point cat g t f b -> Point cat g i f o

Point composition.

Working with isomorphisms.

data Iso cat i o infix 8

An isomorphism is like a Category that works in two directions.

Constructors

Iso infix 8

Fields

fw :: cat i o

bw :: cat o i

inv :: Iso cat i o -> Iso cat o i

Flip an isomorphism.

Specialized lens contexts.

type Total = (->)

Context that represents computations that always produce an output.

type Partial = Kleisli Maybe

Context that represents computations that might silently fail.

type Failing e = Kleisli (Either e)

Context that represents computations that might fail with some error.

Arrow type class for failing with some error.

class Arrow a => ArrowFail e a where

The ArrowFail class is similar to ArrowZero, but additionally embeds some error value in the computation instead of throwing it away.

Minimal complete definition

failArrow

Methods

failArrow :: a e c

Orphan instances

Functor f => Functor (Kleisli f i)
Missing Functor instance for Kleisli.

Methods

fmap :: (a -> b) -> Kleisli f i a -> Kleisli f i b

(<$) :: a -> Kleisli f i b -> Kleisli f i a

Applicative f => Applicative (Kleisli f i)
Missing Applicative instance for Kleisli.

Methods

pure :: a -> Kleisli f i a

(<*>) :: Kleisli f i (a -> b) -> Kleisli f i a -> Kleisli f i b

liftA2 :: (a -> b -> c) -> Kleisli f i a -> Kleisli f i b -> Kleisli f i c

(*>) :: Kleisli f i a -> Kleisli f i b -> Kleisli f i b

(<*) :: Kleisli f i a -> Kleisli f i b -> Kleisli f i a

Alternative f => Alternative (Kleisli f i)
Missing Alternative instance for Kleisli.

Methods

empty :: Kleisli f i a

(<|>) :: Kleisli f i a -> Kleisli f i a -> Kleisli f i a

some :: Kleisli f i a -> Kleisli f i [a]

many :: Kleisli f i a -> Kleisli f i [a]